first few problems: - make dictbatch serialisable - fix client and producer, pass the correct token type in the producer and don't pass anything when subscribing in the client - Receiver didn't have a queue id property
Status:
Writer works just fine and the cache is working fine.
Consumer receives only one message and that's it. The cursor does get called (MoveNext, GetCurrent). It might be a problem related to using the custom token. The last resort would be to stop passing the "key"/"id" in the custom token and revert back to using a simple stream token and pass the ID in the message itself instead (using IObjectWithUniqueId as before)